WIP: seperate weighted_template test
